requirements.txt
argcomplete==1.12.1
asgiref==3.2.10
boto3==1.16.14
botocore==1.19.14
certifi==2020.11.8
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-s3-storage==0.13.4
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
durationpy==0.5
flex==6.14.1
future==0.18.2
hjson==3.0.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpointer==2.0
kappa==0.6.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.4
pip-tools==5.3.1
placebo==0.9.0
PyJWT==1.7.1
pynamodb==4.3.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rfc3987==1.3.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
text-unidecode==1.3
toml==0.10.2
tqdm==4.51.0
troposphere==2.6.3
uritemplate==3.0.1
urllib3==1.25.11
validate-email==1.3
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
